Fix custom cache key mechanism
According to the docs, multi_fetch_fragments supports the shorthand
syntax, `render @items, cache: true`. This is true for boolean values,
but does not support custom cache keys via proc or lambda values. This
is because the shorthand passes cache to @locals rather than @options,
and the search for the callable object is limited to @options.

This commit adds @locals to that search.